Remove openssl dependency
reqwest's default use of nativetls differs from the AWS SDK's default
use of rustls. This brings the two into alignment.

In the future, I may introduce a feature to allow selecting the TLS
library to use. For now, having a single library (and one less
dependency) is a good thing.